==> Address/prajwal/Feb 20/bk 1850/lib/00_remove special characters.ipynb <==
import pandas as pd
import re
import collections

#df = pd.read_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/mn 1850/data/MN_1850_final_output_tesseract.csv')
#df = pd.read_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/bk 1850/address 2/bk_1850_address2_firstrun.csv')
df = pd.read_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/bk 1850/data/BK_1850_final_output_tesseract.csv')


df.columns


#df['Full Address']

"""
for a in range(0, len(df)):
    print(a)
    #df['Full Address'][a] = str(df['Full Address'][a]).replace('Ç','C')
    #df['Full Address'][a] = str(df['Full Address'][a]).replace('√','Y')
    df['Full Address'][a] = str(df['Full Address'][a]).replace('¬',' ')
    df['Full Address'][a] = str(df['Full Address'][a]).replace('•',' ')
    df['Full Address'][a] = str(df['Full Address'][a]).replace('¢','c')
    
    
    # bottom part also commented out
    
    for i in range(0, 10):
        df['Full Address'][a] = str(df['Full Address'][a]).replace('&&','&')
    """

for a in range(0, len(df)):
    if '.5' in str(df['Full Address'][a]):
        print(a)
        df.at[a, 'Full Address'] = df['Full Address'][a].replace('.5','')
    if ',' in str(df['Full Address'][a]):
        df.at[a, 'Full Address'] = df['Full Address'][a].replace(',','')
    if '.' in str(df['Full Address'][a]):
        df.at[a, 'Full Address'] = df['Full Address'][a].replace('.','')
    if '  ' in str(df['Full Address'][a]):
        df.at[a, 'Full Address'] = df['Full Address'][a].replace('  ',' ')

temp_replace_specials = {}
for a in range(0, len(df)):
    temp_str = re.sub('[^A-Za-z0-9]', ' ', str(df['Full Address'][a]))
    temp_str = temp_str.replace('&  &', '&x', 100).replace(' & ', ' ', 100).replace('&x', '&', 100).replace('&  &', '&', 100).replace('  ', ' & ').replace('&  &', '&', 100).replace('  ', ' ', 100)
    temp_str = temp_str.replace('Yh', 'Y h', 100)
    temp_str = temp_str.replace('NYh', 'NY h', 100)
    temp_str = temp_str.replace(' b ', ' h ', 100)
    temp_str = temp_str.replace('YV', 'W', 100)
    temp_str = temp_str.replace('Ayn', ' Av n ', 100)
    temp_str = temp_str.replace('avn', ' Av n ', 100)
    temp_str = temp_str.replace('Avy', ' Av ', 100)
    temp_str = temp_str.replace('.', '', 100)
    #temp_str = temp_str.replace(',', '', 100)
    temp_str = temp_str.replace(',', '', 100)
    temp_str = temp_str.replace('  ', ' ', 100)
    
    for i in range(0, 10):
        try:
            if temp_str.index('h'+str(i))>=0:
                #print(a,i,'yes')
                temp_str = temp_str.replace('h'+str(i), 'h '+str(i), 100)
        except:
            pass
    temp_replace_specials[a] = temp_str

temp_replace_specials[6]

temp_replace_specials[451].split()

#temp_replace_specials[435].split().index('&')

def hasNumbers(inputString):
    return any(char.isdigit() for char in inputString)

part_after_h = {}
part_before_h = {}
part_before_and = {}
part_after_and = {}

for a in range(0, len(df)):
    loc_of_h = -1
    loc_of_and = -1
    try:
        loc_of_h = temp_replace_specials[a].split().index('h')
    except:
        pass
    try:
        loc_of_and = temp_replace_specials[a].split().index('&')
    except:
        pass
    if loc_of_h >= 0:
        part_after_h[a] = ' '.join(temp_replace_specials[a].split()[loc_of_h+1:])
        part_before_h[a] = ' '.join(temp_replace_specials[a].split()[0:loc_of_h])
    
    if loc_of_h < 0 and loc_of_and >= 0:
        temp_part_before_and = ' '.join(temp_replace_specials[a].split()[0:loc_of_and])
        if hasNumbers(temp_part_before_and):
            part_before_and[a] = ' '.join(temp_replace_specials[a].split()[0:loc_of_and])
            part_after_and[a] = ' '.join(temp_replace_specials[a].split()[loc_of_and+1:])

part_before_and

df = df.rename({'Full Address': 'Full Address old'}, axis=1)

#part_after_h[4]

new_full_address = {}
for a in range(0, len(df)):
    try:
        new_full_address[a] = part_after_h[a]
    except:
        try:
            new_full_address[a] = part_before_and[a]
        except:
            new_full_address[a] = temp_replace_specials[a]

for a in range(0, len(df)):
    split_list_from_address = new_full_address[a].split()
    try:
        if split_list_from_address[-1] == '&':
            split_list_from_address = split_list_from_address[:-1:]
    except:
        pass
    try:
        if split_list_from_address[-1] == 'n':
            split_list_from_address = split_list_from_address[:-1:]
    except:
        pass
    try:
        if split_list_from_address[0] == '&':
            split_list_from_address = split_list_from_address[1::]
    except:
        pass
    new_full_address[a] = ' '.join(split_list_from_address)

def firstUpperCaseLetter(str) : 
    for i in range(0, len(str)) : 
        if (str[i].istitle()) : 
            return i
    return 0

for a in range(0, len(df)):
    test_string1 = new_full_address[a]
    if hasNumbers(test_string1)==False:
        loc_of_first_uppercase_letter = firstUpperCaseLetter(test_string1)
        new_full_address[a] = test_string1[loc_of_first_uppercase_letter:]

new_full_address[996]

for a in range(0, len(df)):
    try:
        if new_full_address[a].index(' and ')>=0:
            print(a, new_full_address[a])
    except:
        pass

new_full_address

#temp_replace_specials[497]

df['Full Address'] = new_full_address.values()

for a in range(0, len(df)):
    try:
        if len(part_after_and[a])>0:
            pass
    except:
        part_after_and[a] = ''

for a in range(0, len(df)):
    try:
        if len(part_before_h[a])>0:
            pass
    except:
        part_before_h[a] = ''

part_after_and_complete = collections.OrderedDict(sorted(part_after_and.items()))
part_before_h_complete = collections.OrderedDict(sorted(part_before_h.items()))

df['part after and'] = list(part_after_and_complete.values())
df['part before h'] = list(part_before_h_complete.values())

#df.to_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/mn 1850/data/MN_1850 tesseract spl removed.csv')
df.to_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/bk 1850/data/bk_1850_spl_removed.csv')
