==> Address/prajwal/Feb 20/bk 1850/lib/01.5_code for address 2.ipynb <==
import numpy as np
import pandas as pd

df = pd.read_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/bk 1850/data/bk_1850_spl_removed.csv')

df.columns

address2 = []
for i in range(0, len(df)):
    str1 = str(df['part after and'][i]).replace('nan','')
    str2 = str(df['part before h'][i]).replace('nan','')
    str3 = str(df['Address 2'][i]).replace('nan','')
    str4 = (str1 + ' ' + str2 + ' ' + str3).replace('  ',' ').replace('  ',' ').replace('  ',' ')
    address2.append(str4)


for i in range(0,len(df)):
    try:
        if address2[i].split()[-1]=='N &':
            address2[i] = ''.join(address2[i][:-4])
    except:
        pass
    try:
        if address2[i].split()[0]=='&':
            address2[i] = ''.join(address2[i][1:])
    except:
        pass
    try:
        if address2[i].split()[-1]=='&':
            address2[i] = ''.join(address2[i][:-2])
    except:
        pass
    try:
        if address2[i].split()[-1]=='N':
            address2[i] = ''.join(address2[i][:-2])
    except:
        pass
    address2[i] = address2[i].replace('N Y','').replace('NY','')

df2 = pd.DataFrame(address2)
df2.columns = ['Full Address']

df2.to_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/bk 1850/data/bk_1850_input_for_address2.csv')
