==> Address/prajwal/Feb 20/bk 1850/lib/11_Address Builder copy.Rmd <==
---
title: "Address Builder"
author: "Clinton"
date: "26 February 2020"
output: html_document
---

### 1. Set Up

```{r message = FALSE, warning = FALSE}
library(dplyr)
library(stringr)
library(english)
library(purrr)
library(knitr)
library(kableExtra)

mn_or_bk <- 'mn'
year <- '1880'
doubttype <- 'not so confident' #doubt or not so confident

#mn_output <- readRDS("/Users/prajwal/Desktop/Columbia/C4SR/hnyc_census-master/mn_output.rds")
#write.csv(mn_output, '/Users/prajwal/Desktop/Columbia/C4SR/bk_1850/mn_output.csv')
input1 = paste('/Users/prajwal/Desktop/Columbia/C4SR/',mn_or_bk,'_',year,'/for address builder/',mn_or_bk,'_',year,'_',doubttype,'_ready4.csv',sep='')


#mn_output <- read.csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract output/run1/after cleaning/mn-1880-for-address-builder.csv')
#mn_output <- read.csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract output/run2/after cleaning/bk-1850-for-address-builder.csv')
#mn_output <- read.csv(input1)
#mn_output <- read.csv('/Users/prajwal/Desktop/Columbia/C4SR/bk_1850/for address builder/bk_1850_not so confident_ready3.csv')
#mn_output <- read.csv('/Users/prajwal/Desktop/Columbia/C4SR/bk_1850/after cleaning/temp3.csv')

#mn_output <- read.csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/mn 1850/data/mn-1850-for-address-builder.csv')
#mn_output <- read.csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/mn 1850/data/mn_1850_for_rerun_cleaned.csv')
mn_output <- read.csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/bk 1850/address 2/data2/bk-1850-3-address2-for-rerun-cleaned.csv')
```

<br>

### 2. Street Builder

```{r}
# Street type builder
street_type_builder = function(df) {
  df$street_type = case_when(
    str_detect(df$best_match, pattern = " ST") ~ "ST",
    str_detect(df$best_match, pattern = " DR") ~ "DR",
    str_detect(df$best_match, pattern = " CIR") ~ "CIR",
    str_detect(df$best_match, pattern = " AVE") ~ "AVE",
    str_detect(df$best_match, pattern = " PL") ~ "PL",
    str_detect(df$best_match, pattern = " CT") ~ "CT",
    str_detect(df$best_match, pattern = " PARK") ~ "PARK",
    str_detect(df$best_match, pattern = " PLZ") ~ "PLZ",
    str_detect(df$best_match, pattern = " PKWY") ~ "PKWY",
    str_detect(df$best_match, pattern = " WAY") ~ "WAY",
    str_detect(df$best_match, pattern = " ALY") ~ "ALY",
    str_detect(df$best_match, pattern = " PIER") ~ "PIER",
    str_detect(df$best_match, pattern = "PIER") ~ "PIER",
    str_detect(df$best_match, pattern = " SLIP") ~ "SLIP",
    str_detect(df$best_match, pattern = " ROW") ~ "ROW",
    str_detect(df$best_match, pattern = " APPROACH") ~ "APPROACH",
    str_detect(df$best_match, pattern = " LN") ~ "LN",
    str_detect(df$best_match, pattern = " TER") ~ "TER",
    str_detect(df$best_match, pattern = " HTS") ~ "HTS",
    str_detect(df$best_match, pattern = " BLVD") ~ "BLVD",
    str_detect(df$best_match, pattern = " BRG") ~ "BRG",
    str_detect(df$best_match, pattern = " HL") ~ "HL",
    str_detect(df$best_match, pattern = "AVE") ~ "AVE",
    str_detect(df$best_match, pattern = "BROADWAY") ~ "AVE",
    TRUE ~ "ST")
  
  return(df)
}

mn_output = street_type_builder(mn_output)
```

<br>

### 3. Address Builder

The function takes three variables `hn_1`, `best_match` and `street_type`, and concatenates an address suitable for geocoding. Two parameters are available for customising the output address. 

* The first parameter `spell_ave` defaults to FALSE, and converts numerics to words for streets where `street_type` is avenue. For example, "3 AVE" will be transformed to "THIRD AVE".

* The second parameter `suffix` defaults to FALSE, and adds ordinal indicators (e.g. st, nd, rd) for all streets. For example, "108 ST" will be transformed to "108TH ST".

```{r}
build_Address = function(df, spell_ave = FALSE, suffix = FALSE) {
  
  # Variables used for Concatenate, replacing with empty string in order to sanitise input for functions below. Also because otherwise NA would be concatenated
  print(df$best_match)
  tmp_hn              = ifelse(!is.na(df$hn_1),
                               df$hn_1,
                               "")
  tmp_best_match      = df$best_match
  tmp_street_type     = ifelse(!is.na(df$street_type),
                               df$street_type,
                               "")
  
  # # Spell Avenue numerics if TRUE
  # if (spell_ave == TRUE) {
  #   
  #   tmp_hn = ifelse(str_detect(tmp_street_type, 
  #                              pattern = "AVE"),
  #                   as.character(as.english(df$hn_1)),
  #                   tmp_hn)
  # }
  # # Adding suffixes to numerics if TRUE
  if (suffix == TRUE) {
    
    tmp_ordinal = str_extract(tmp_best_match, 
                              pattern = "[:digit:]+") %>%
      as.integer() %>%
      map_chr(., 
              function(x) {
                ifelse(!is.na(x),
                       scales::ordinal(x),
                       "")
              })

    tmp_best_match = str_replace(tmp_best_match, 
                                 pattern = "[:digit:]+",
                                 tmp_ordinal)
  }
  # Removing duplicated street type if street type already in best match for street types that are not "ST"
  tmp_street_type = ifelse(str_detect(tmp_best_match,
                                      pattern = paste0(tmp_street_type, "$")),
                           "",
                           tmp_street_type)
  
  # Concatenating Address
  addr = paste(tmp_hn, 
               tmp_best_match, 
               tmp_street_type) %>%
    str_trim(side = "both") %>%
    toupper()
  
  # Replacing empty strings back with NA
  addr = ifelse(addr == "" | str_detect(addr, pattern = "^ST$"),
                NA_character_,
                addr)
  # Appending to data frame
  df$addr = addr
  
  return(df)
}
```

<br>

### 3. Examples

```{r}
set.seed(2468)
#index = sample(NROW(mn_output), 100)
#index = sample(NROW(mn_output), 3)
#examples = mn_output[index,]
examples = mn_output

base_function   = build_Address(examples)$addr
#spell_ave_true  = build_Address(examples, spell_ave = TRUE)$addr
suffix_true     = build_Address(examples, suffix = TRUE)$addr
both_true       = build_Address(examples, spell_ave = TRUE, suffix = TRUE)$addr

# data.frame(hn_1 = examples$hn_1,
#            best_match = examples$best_match,
#            street_type = examples$street_type, 
#            base_function, 
#            spell_ave_true,
#            suffix_true,
#            both_true) %>%
#   kable() %>%
#   kable_styling() %>%
#   scroll_box(width = "100%", height = "300px")

data.frame(hn_1 = examples$hn_1,
           best_match = examples$best_match,
           street_type = examples$street_type, 
           base_function, 
           suffix_true,
           both_true) %>%
  kable() %>%
  kable_styling() %>%
  scroll_box(width = "100%", height = "300px")


# save_this <- data.frame(hn_1 = examples$hn_1,
#            best_match = examples$best_match,
#            street_type = examples$street_type, 
#            base_function, 
#            spell_ave_true,
#            suffix_true,
#            both_true)

save_this <- data.frame(hn_1 = examples$hn_1,
           best_match = examples$best_match,
           street_type = examples$street_type, 
           base_function, 
           suffix_true,
           both_true)


output1 = paste('/Users/prajwal/Desktop/Columbia/C4SR/',mn_or_bk,'_',year,'/after cleaning/',mn_or_bk,'_',year,'_',doubttype,'_built.csv',sep='')

#write.csv(save_this, output1)
#write.csv(save_this,'/Users/prajwal/Desktop/Columbia/C4SR/tesseract output/run1/final/mn_1880_final_tesseract.csv')
#write.csv(save_this,'/Users/prajwal/Desktop/Columbia/C4SR/tesseract output/run2/after cleaning/bk_1850_final_tesseract_run2.csv')
#write.csv(save_this,'/Users/prajwal/Desktop/Columbia/C4SR/tesseract output/run2/for rerun/bk_1850_final_after_rerun.csv')

#write.csv(save_this,'/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/mn 1850/data/mn-1850-final_tesseract.csv')
#write.csv(save_this,'/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/mn 1850/data/mn-1850-final_after_rereun.csv')
write.csv(save_this,'/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/bk 1850/address 2/data2/bk-1850-3-address2-final.csv')

```

<br>

### 4. Future Improvements

* Function works as expected but code could be improved and cut down to be more efficient

***
